LayoutTests/http/tests/inspector/network/network-empty-xhr-expected.txt

CONSOLE MESSAGE: line 18: Done.
Tests empty xhr content is correctly loaded in inspector.

Bug 79026
http://127.0.0.1:8000/inspector/network/resources/empty.html
resource.content: 
http://127.0.0.1:8000/inspector/network/resources/empty.html
resource.content: 
